fix: :ambulance: Allow `Kolkra.owner_ids` to be set in config; fix how reading native bans is handled
I assumed that `Guild.fetch_ban` would return an optional, but apparently not--if the user we're looking up isn't banned, it throws a `NotFound` exception. D'oh! I wrote a tiny helper to catch that and return `None` instead to avoid a major refactor.

I also added a yellow accent color to the `ModWarning` class's embed bases that I seem to have forgotten during initial development.